Scaffold the build interface
The `Builder` struct defines the public interface for building a PGXN
release. Internally, it delegates `configure`, `compile`, and `test`
commands (so far) to pipeline-specific build modules. Its `new`
constructor reads the pgxn_meta Release pipeline to determine which one
to instantiate; in the future it will also fall back on detecting the
pipeline where none is specified.

The `error` module define errors returned by the crate, while the
`pipeline` module defines a trait with the functions `Builder` expects
to dispatch to. All pipelines must implement this trait. This will make
it easy to add new pipelines in the future; They just have to implement
the trait, add a new `Build` enum, then add appropriate calls to its
methods in `Builder`'s dispatching methods.

For now, the `pgxs` and `pgrx` modules alone define structs that
implement the `pipeline.Pipeline` trait with no-op functions.

Add tests for basic functionality of `Builder`.