Implement template fetching and execution
PGXN predates provides a file, `index.json`, at the root of every mirror
that maps API names to RFC 6570 templates. The correct way to call the
APIs is to load this file and use the templates to generate the paths to
APIs relative to a base URL.

Add that capability to the api module. Use [iri_string::template] to
parse and evaluate the templates and store them in the struct. Add
functions to fetch and load file contents and JSON files, as this will
be a common function, and use it to fetch `index.json`. Details:

*   `parse_base_url()` parses the base URL, ensuring that it always ends
    in a slash. Save the resulting URL instead of a string.
*   `get_file` opens a file and returns the file handle, and customizes
    errors to include the file name.
*   `fetch_json` fetches the JSON from a URL and returns the resulting
    serde_json::Value. For `file:` URLs it uses `get_file` to read the
    file; for `http` and `https` URLs it uses the agent passed to it. It
    allows `http` for tests purposes; the agent created by `new` always
    requires `https`, so this should be safe.
*   `type_of` returns the type of a serde_json::Value. Used for error
    messages.
*   `fetch_templates` uses `fetch_json` to fetch an `index.json` and
    parses it into a HashMap of template names and UriTemplateString
    values.
*   `url_for` builds a URL for from a template and returns it relative
    to the base URL.

Tested as fully as I can determine; test coverage is currently broken;
will have to return later to fill the gaps.

While at it:

*   Use the cargo package name and version for the user agent string.
*   Customize IO errors to include more contextual information,
    including the action that triggered an error and, where possible,
    the file name or URL.
*   Update `download_to` to use the `download` template to build the URL
    to download.
*   Make the arguments to `new` be references; it doesn't retain them
    and used them entirely as references, anyway.

  [iri_string::template]: https://docs.rs/iri-string/latest/iri_string/template/index.html