Implement the meta API
Sadly, the current PGXN manager strips the meta-spec out of this file. I
guess because I had thought back in 2011 or whenever that the API itself
was version enough. Whatever.

So inject a meta-api version if there isn't one. Then convert it to a
pgxn_meta::release::Release struct. For now, convert errors to strings;
the next release of pgxn_meta will have more reasonable errors we can
properly wrap.