Upgrade pgxn_meta and add download validation
Take advantage of the errors defined in pgxn_meta v0.5.0 rather than
stringifying them. Add tests for errors returned by the `meta` and
`download` APIs.

Change the signature for `download_to` to take a pgxn_meta Release
struct instead of a distribution name and version. Then use the Release
both to determine the name and version to download, and then to validate
the download against the digests in the Release. Update the tests for
the new signature, and to ensure digest validation.

Also: Update dependencies and fill out the Cargo.toml `package` details.