Cover overlooked branches
I got test coverage fixed in my Mac, so could see what I've missed in
the last several commits. Fill those gaps as best as possible. There is
only one branch I can't figure out how to trigger: copying from an http
request to a file. Perhaps in the future we'll update the Api struct to
support dependency injection, or switch to some more mockable request
agent. But in the meantime, use a macro to reduce that branch to a
single line.

Other changes: Eliminate an error check where the url docs say there
will never be an error, add a `clean` target to the `Makefile`, and
convert `type_of` to a macro.