Pass build directory to pipelines
Instead of the release metadata. The build pipelines don't care about
the metadata, but need to know the directory in which to execute. Also
add a placeholder for using `sudo`.

Use a `PathBuf` in the pipeline implementations, rather than
`AsRef<Path>`, mostly because I couldn't figure out the proper
incantation to hold onto an `AsRef` through a trait definition. Not
worth the time, so just convert to a `PathBuf` and worry about it later
(if ever).

Hold onto the metadata in the parent method, however, as it may well
need to use it at some point. This requires a bit of futzing while
looking for the pipeline type in the metadata. Will implement the
pipeline-detection `todo!()`s next to address it.

While at it, remove the `download` and `unpack` methods from `Builder`,
as they're implemented by `api::Api`.