Implement command execution and PGXS configuration
Add `run` to the `Pipeline` trait with a full implementation that
executes a command. It takes a `sudo` param to use execute the command
with elevated privileges, though it's currently ignored on Windows (as
copiously covered in comments).

Implement the basic behavior for configuring, compiling, testing, and
installing PGXS extensions, and fully test the `configure`
implementation. Note that the tests actually execute a `configure` shell
script. This works on the GitHub Windows runner, too, which have Bash
installed.

In the process, return to storing the directory in the pipeline
implementations and setting them on the Command object, because then the
scope of the directory-setting is in the command, and not the parent
process. Otherwise we'd have to get up to a bunch of shenanigans to
change to the working directory in a thread-safe way. Fortunately, I
finally worked out the pattern to store the directory as an
`AsRef<Path>` everywhere, which simplifies usability quite a bit (at the
expense of a bunch of annotations).

Other changes:

*   Upgrade the `url` crate, since 2.5.3 has been yanked (resolves #13).
*   Adopt the `assertables` crate for nicer test assertions. Will be
    used more in the future.
*   Move `#[cfg(test)]` to the bottom of module files, because
    `--excl-start` in `.ci/test-cover` excludes everything after that
    declaration, not just the module. As a result, add a test for no
    release data to `api/dist`.