Add PgConfig module
It takes a path to `pg_config`, executes it, and stores all the
configuration values. A `get` method provides access to configurations
(using a lowercase key name), while `iter` and `into_iter` allow
iteration over all the configurations. The module is crate-only for now,
and not yet used, but will be soon.

In order to support tests on multiple OSes, add `mocks/pg_config.rs` and
`mocks/exit_err.rs`. The PgConfig tests use `rustc` to compile apps for
the current OS that mock the output of `pg_config` and that exit with an
error. This eliminates the need for a bunch of conditional code to build
shell scripts or batch files; we're testing Rust, so Rust is available,
so just use it to build the apps and ensure consistent behavior.

Tweak the handling of Command errors by stringifying a Command before
creating the error object. This avoids a bunch of ownership issues. The
code does the work only in conditional blocks that execute when an error
is needed, so the overhead is minimal.