Use rust to mock CLI apps in tests
Already done in the pg_config tests, but adapted for all the other tests
that actually call CLIs. Move `compile_mock` to `lib/tests.rs` to use
throughout tests, use the new `mocks/echo.rs` to mock `sudo` and other
commands, and remove all the code building conditional shell or batch
scripts. The exception is `configure`, which has to be a shell script.
Also remove the comments about `sudo` on Windows, since the tests now
successfully mock it.

Add code to check Command success, too. This was previously lacking,
leading to execution that succeeded because an app exists (e.g. `make`)
and therefore successfully executed but exited with an error. For that
case, check `status.success` on command output and pass `stderr` to the
resulting error if not successful. Fix a bug thus found where
`mocks/exit_err.rs` was printing to STDOUT instead of STDERR. Also fix
PGXS tests that were previously successfully executing but should not
have been.