Add `is_writeable` to the pipeline trait
It has an implementation, and will be used to determine whether `sudo`
is likely to be needed to install an extension.

It needs to stringify file names, so replace the `filename!` macro in
the `api` module with a function in the root crate and make it public
inside the create, with some tests.