Replace `sudo` param with `maybe_sudo` method.
The `maybe_sudo` method creates a new command based on whether `sudo`
might be needed and, if so, whether `pkglibdir` is writable. If `sudo`
is true and `pkglibdir` is not writable it creates a Command with `sudo`
as the program and the program passed as its first argument. Otherwise
it returns a Command with the passed program.

With that intelligence, remove the `sudo` param from the interface and
all methods. There will be no need to specify `sudo` as a boolean, as
the package can figure it out for itself.

In the future it might be necessary to allow the user to specify a path
to `sudo` (and `make` and `cargo` and whatever else?), but for now it
will be enough to have the user make sure that such commands are in the
path.

Also remove `mut` from `PgConfig::get`, as it's unnecessary and makes
its usage more difficult. Found while using it in `maybe_sudo`.