Add logging and capture  & style execution output
Add extensive info, debug, and trace logging throughout. The CLI will
style and emit log messages at the INFO level or higher.

Upgrade pgxn_meta to v0.6.0 and add a test to the API to ensure it can
properly handle Meta API requests to api.pgxn.org, which adds fields
supported by pgxn_meta v0.6.0.

Add the `lines` module, which defines a trait, `line::WriteLine`, for
writing lines of text. It also defines two implementations:
`LineWriter`, which simply writes lines to a `std::io::Write`, and
`ColorLine`, which uses an `owo_colors::Style` to style a line before
writing it to a `std::io::Write`. Use `LineWriter` to capture output
during tests, so it doesn't appear in test output. Hopefully we'll also
eventually be able to test that output, but for now ownership rules
prevent it.

Add the `exec` module to set the directory context for the execution of
a command and to stream its STDOUT and STDERR output to
`lines::WriteLine` values. This allows for the styling and capturing of
output.

Import the `api` and `pg_config` modules instead of making them directly
public.

Remove the generics and lifetimes from most structs. This greatly
simplifies building structs from other structs, and allows the
`exec::Executor` to be passed to Pipeline constructors instead of a
directory. This allows the module to set up the default execution output
streaming, styling STDOUT in light grey and STDERR in red by streaming
them to `line::ColorLine` structs that wrap our STDOUT and STDERR.
Ultimately this configuration may move to consumers.

For better or worse, the need for the line writers to be mutable
requires that the Pipelines be mutable, too.